Extract methods for finding languages
